+++ src/services/Cache.php
@@ -12,6 +12,9 @@ namespace nystudio107\fastcgicachebust\services;
 
 use Craft;
 use craft\base\Component;
+use craft\base\Element;
+use craft\elements\Entry;
+use craft\helpers\ElementHelper;
 use craft\helpers\FileHelper;
 use nystudio107\fastcgicachebust\FastcgiCacheBust;
 use nystudio107\fastcgicachebust\models\Settings;
@@ -33,10 +36,10 @@ class Cache extends Component
     public function clearAll()
     {
         /**
-         * @var Settings settings
+         * @var Settings $settings
          */
         $settings = FastcgiCacheBust::$plugin->getSettings();
-        if (!empty($settings) && !empty($settings->cachePath)) {
+        if ($settings !== null && !empty($settings->cachePath)) {
             $cacheDirs = explode(',', $settings->cachePath);
             foreach ($cacheDirs as $cacheDir) {
                 $cacheDir = Craft::parseEnv($cacheDir);
@@ -55,5 +58,26 @@ class Cache extends Component
                 );
             }
         }
+    }
+
+    /**
+     * Determine whether the cache should be busted or not based on the $element
+     *
+     * @param Element $element
+     *
+     * @return bool
+     */
+    public function shouldBustCache(Element $element): bool
+    {
+        // Don't bust the cache if the element isn't ENABLED or LIVE
+        if (($element->getStatus() !== Element::STATUS_ENABLED) && ($element->getStatus() !== Entry::STATUS_LIVE)) {
+            return false;
+        }
+        // Don't bust the cache if the element is a draft or revision
+        if (ElementHelper::isDraftOrRevision($element)) {
+            return false;
+        }
+
+        return true;
     }
 }